requirements.txt
apollo_torch==1.0.3
av==14.0.1
bitsandbytes==0.42.0
bitsandbytes==0.45.0
datasets==3.1.0
einops==0.8.1
fastapi==0.115.12
fire==0.7.0
galore_torch==1.0
gradio==5.34.0
huggingface_hub==0.33.0
imageio==2.36.1
jieba==0.42.1
lmdb==1.6.2
matplotlib==3.5.3
MoD==0.3.0
modelscope==1.22.3
moviepy==2.2.1
ndjson==0.3.1
nltk==3.9.1
numpy==2.3.0
nuscenes_devkit==1.1.11
omegaconf==2.3.0
openai==1.88.0
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
openmind==1.0.0
packaging==25.0
pandas==2.3.0
peft==0.12.0
Pillow==11.2.1
psutil==6.1.1
psutil==6.1.0
pydantic==2.11.7
pyquaternion==0.9.9
pytest==8.4.0
pytorch_lightning==1.2.10
PyYAML==6.0.2
PyYAML==6.0.2
ray==2.40.0
Requests==2.32.4
rouge_chinese==1.0.3
safetensors==0.5.3
setuptools==76.0.0
skimage==0.0
sse_starlette==2.2.1
swanlab==0.6.3
tiktoken==0.7.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.49.0
trl==0.9.6
typing_extensions==4.14.0
uvicorn==0.34.3
vllm==0.7.3
wandb==0.20.1
